[breaking][rename] `*Parser*` to `*Runnable*`
While parsing is definitely a responsibility of this class hierarchy, (for now) that task is always discharged in the same way i.e. with `ast.parse().` On the other hand, massaging the code (like adding implicit `import`s, wrapping an imperative script in a function etc.) is polymorphic w.r.t. the kind of user script to run.

While we are at it, fix a few typos in docstrings.